refactor: module for registry interfaces and events
Introduces the module 'registty'

This module contains the `EventListenerEvents`: `EventListenerRegistered`
and `EventListenerDeregistered`

and

the module 'socket' after the `Socket` interface on the `Registry` daemon.

Where we'll find the `Available` event and its impls.

 Conflicts:
	atspi-common/src/events/mod.rs
